Moving things into new directories to help make pull easier
