+++ src/main/java/tempest/ui/ErrorMessage.java
@@ -1,4 +1,4 @@
-package tempest.GUI;
+package tempest.ui;
 
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
