+++ src/main/java/tempest/ui/components/ClearButton.java
@@ -1,8 +1,8 @@
-package tempest.GUI.components;
+package tempest.ui.components;
 
 import javax.swing.JButton;
 
-import tempest.GUI.Page;
+import tempest.ui.pages.Page;
 
 public class ClearButton extends JButton {
     private static final long serialVersionUID = -857346759811441764L;
