+++ src/main/java/tempest/ui/pages/AddModulePage.java
@@ -1,4 +1,4 @@
-package tempest.GUI;
+package tempest.ui.pages;
 
 import java.awt.event.ActionEvent;
 
@@ -6,9 +6,10 @@ import javax.swing.*;
 
 import tempest.Module;
 import tempest.State;
-import tempest.GUI.components.ActionButtonPanel;
-import tempest.GUI.components.ClearButton;
-import tempest.GUI.components.ModuleDropDown;
+import tempest.ui.*;
+import tempest.ui.components.ActionButtonPanel;
+import tempest.ui.components.ClearButton;
+import tempest.ui.components.ModuleDropDown;
 
 public class AddModulePage extends Page {
     private final State state;
