+++ src/test/java/tempest/GUI/GUIManagerTest.java
@@ -7,6 +7,11 @@ import org.junit.Test;
 
 import tempest.State;
 import tempest.Supervisor;
+import tempest.ui.ErrorMessage;
+import tempest.ui.GUIManager;
+import tempest.ui.pages.HomePage;
+import tempest.ui.pages.ManageModulesPage;
+import tempest.ui.pages.ManageSessionsPage;
 
 public class GUIManagerTest {
     State state = new State();
